Move `move` to filesystem utility
